added test cases
